requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.6
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
openai==1.64.0
httpx==0.28.1
svix==1.24.0
